fix: Adding all metadata for data set report
