--- src/components/data-set-report/data-set-report.html
@@ -1,35 +1,0 @@
-<ion-row>
-  <ion-col>
-    <div *ngIf="isLoading">
-      <loading [loadingMessage]="loadingMessage" [loadingSize]="'large'"></loading>
-    </div>
-    <div *ngIf="!isLoading">
-      <div *ngIf="entryFormSections && entryFormSections.length > 0">
-        <div *ngFor="let section of entryFormSections;trackBy: trackByFn">
-          <table width="100%">
-            <tbody>
-              <tr *ngIf="section.name != ''">
-                <td [ngStyle]="{'background-color': '#F4F4F4'}">
-                  {{ section.name }}
-                </td>
-              </tr>
-              <tr *ngIf="section.dataElements">
-                <td>
-                  <div text-left>
-                    <div *ngFor="let dataElement of section.dataElements;trackBy: trackByFn">
-                      <data-set-report-row [dataElement]="dataElement" [dataValues]="dataElementDataValuesMapper[dataElement.id]" [appSettings]="appSettings"></data-set-report-row>
-                    </div>
-                  </div>
-                </td>
-              </tr>
-            </tbody>
-          </table>
-        </div>
-      </div>
-      <div *ngIf="entryFormSections && entryFormSections.length == 0">
-        <empty-list-notification [emptyListMessage]="translationMapper['Missing entry form sections, please contact administarator']"></empty-list-notification>
-      </div>
-
-    </div>
-  </ion-col>
-</ion-row>
