+++ src/pages/reports/components/data-set-report/data-set-report.html
@@ -1,0 +1,42 @@
+<ion-row>
+  <ion-col>
+    <div *ngIf="isLoading">
+      <loading [loadingMessage]="loadingMessage" [loadingSize]="'large'"></loading>
+    </div>
+    <div *ngIf="!isLoading">
+      <div *ngIf="entryFormType && entryFormType == 'CUSTOM'">
+        <custom-data-entry-form [dataEntryFormDesign]="dataEntryFormDesign" [data]="{}" [entryFormType]="'aggregate'"
+          [entryFormSections]="entryFormSections"></custom-data-entry-form>
+      </div>
+      <div *ngIf="entryFormType && entryFormType != 'CUSTOM'">
+        <div *ngIf="entryFormSections && entryFormSections.length > 0">
+          <div *ngFor="let section of entryFormSections;trackBy: trackByFn">
+            <table width="100%">
+              <tbody>
+                <tr *ngIf="section.name != ''">
+                  <td [ngStyle]="{'background-color': '#F4F4F4'}">
+                    {{ section.name }}
+                  </td>
+                </tr>
+                <tr *ngIf="section.dataElements">
+                  <td>
+                    <div text-left>
+                      <div *ngFor="let dataElement of section.dataElements;trackBy: trackByFn">
+                        <data-set-report-row [dataElement]="dataElement" [dataSetReportAggregateValues]="dataSetReportAggregateValues"
+                          [appSettings]="appSettings"></data-set-report-row>
+                      </div>
+                    </div>
+                  </td>
+                </tr>
+              </tbody>
+            </table>
+          </div>
+        </div>
+      </div>
+      <div *ngIf="entryFormSections && entryFormSections.length == 0">
+        <empty-list-notification [emptyListMessage]="'Missing entry form sections, please contact administarator'"></empty-list-notification>
+      </div>
+
+    </div>
+  </ion-col>
+</ion-row>
