+++ src/pages/sync/components/aggregate-data-downloader/aggregate-data-downloader.ts
@@ -182,7 +182,7 @@ export class AggregateDataDownloaderComponent implements OnInit {
           this.progressTrackerPacentage = 100;
           setTimeout(() => {
             this.showLoader = false;
-          }, 50);
+          }, 100);
           this.appProvider.setTopNotification(
             `Discovered aggregate data has been saved successfully`
           );
