+++ src/pages/sync/components/event-data-downloader/event-data-downloader.ts
@@ -131,7 +131,7 @@ export class EventDataDownloaderComponent implements OnInit {
                   this.progressTrackerPacentage = 100;
                   setTimeout(() => {
                     this.showLoader = false;
-                  }, 50);
+                  }, 100);
                   this.appProvider.setTopNotification(
                     `Discovered events data has been saved successfully`
                   );
